+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Bump all dependencies to the latest versions
 - Apply linter suggestions to the project
+- Big refactoring remove duplicated code and prepare to start adding tests
 
 ## [1.7.1] = 2021-05-31